+++ src/three/gltf/MeshFeatures.js
@@ -143,7 +143,7 @@ export class MeshFeatures {
 
 		} else {
 
-			return Promise.resolve().then( () => {
+			return queueMicrotask( () => {
 
 				return this.getFeatures( ...args );